Cleaned up script to comply with shellcheck

+++ DotfileSetup.sh
@@ -38,9 +38,8 @@ echo "Creating ~/dotfile-backup..."
 # Use dbd (dotfile backup directory) so we can create new backups per-date
 # Honestly, this is my first time trying a script like this, I have no idea
 # what dumb accidents might happen. It appears to work on my machine, at least.
-dbd=~/dotfile-backup/$(date +%y-%m-%d_%T)_dotfiles/
 # If date fails somehow, just default to ~/dotfile-backup
-if [ $? -ne 0 ]; then
+if ! dbd=~/dotfile-backup/$(date +%y-%m-%d_%T)_dotfiles/; then
   dbd=~/dotfile-backup/
 fi
 
@@ -50,14 +49,14 @@ echo "Moving files from $HOME to $dbd..."
 
 # Move files and folders, ignoring $dfp/config and the setup
 # to newly created $dbd
-for FILENAME in $(ls $dfp | grep -v config | grep -v Dotfile)
+for FILENAME in $(ls "$dfp" | grep -v config | grep -v Dotfile)
 do
-  if [ -h ~/.$FILENAME ]; then
-    echo "Removing symlink .$FILENAME"
-    rm ~/.$FILENAME
-  elif [ -a ~/.$FILENAME ]; then
+  if [ -h "$HOME/.$FILENAME" ]; then
+    echo "Removing symlink $HOME/.$FILENAME"
+    rm "$HOME/.$FILENAME"
+  elif [ -a "$HOME/.$FILENAME" ]; then
     echo "Moving .$FILENAME"
-    mv ~/.$FILENAME $dbd
+    mv "$HOME/.$FILENAME" "$dbd"
   fi
 done
 
@@ -66,72 +65,79 @@ echo "Moving folders from $XDG_CONFIG_HOME to $dbd.config"
 
 # Move files from XDG_CONFIG_HOME to $dbd/.config
 # This is probably terrible in some way I don't understand.
-for FILENAME in $(ls $dfp/config)
+for FILENAME in $(ls "$dfp/config")
 do
-  if [ -h $XDG_CONFIG_HOME/$FILENAME ]; then
+  if [ -h "$XDG_CONFIG_HOME/$FILENAME" ]; then
     echo "Removing symlink $XDG_CONFIG_HOME/$FILENAME"
-    rm $XDG_CONFIG_HOME/$FILENAME
-  elif [ -a $XDG_CONFIG_HOME/$FILENAME ]; then
+    rm "$XDG_CONFIG_HOME/$FILENAME"
+  elif [ -a "$XDG_CONFIG_HOME/$FILENAME" ]; then
     echo "Moving $XDG_CONFIG_HOME/$FILENAME"
-    mv $XDG_CONFIG_HOME/$FILENAME $dbd/.config/
+    mv "$XDG_CONFIG_HOME/$FILENAME" $dbd/.config/
   fi
 done
 
 # If nothing got copied to the backup dir, go ahead and delete it.
-if [ $(du -B 1 $dbd | cut -f 1 | tail -n 1) -le 8192 ]; then
+if [ "$(du -B 1 $dbd | cut -f 1 | tail -n 1)" -le 8192 ]; then
   rm -rf $dbd
   echo
   echo "No backups made! Cleaning up ~/dotfile-backup..."
 fi
 
 # Same, but for the whole backup dir
-if [ $(du -B 1 ~/dotfile-backup | cut -f 1 | tail -n 1) -le 8192 ]; then
+if [ "$(du -B 1 ~/dotfile-backup | cut -f 1 | tail -n 1)" -le 8192 ]; then
   rm -rf ~/dotfile-backup
 fi
 
 echo
 echo "Linking..."
 
-for FILENAME in $(ls $dfp | grep -v config | grep -v Dotfile)
+for FILENAME in $(ls "$dfp" | grep -v config | grep -v Dotfile)
 do
   echo "Linking $FILENAME to $HOME/.$FILENAME"
-  ln -s $dfp/$FILENAME $HOME/.$FILENAME
+  ln -s "$dfp/$FILENAME" "$HOME/.$FILENAME"
 done
 
-mkdir -p $XDG_CONFIG_HOME
+mkdir -p "$XDG_CONFIG_HOME"
 
-for FILENAME in $(ls $dfp/config)
+for FILENAME in $(ls "$dfp/config")
 do
   echo "Linking config/$FILENAME to $XDG_CONFIG_HOME/$FILENAME"
-  ln -s $dfp/config/$FILENAME $XDG_CONFIG_HOME/$FILENAME
+  ln -s "$dfp/config/$FILENAME" "$XDG_CONFIG_HOME/$FILENAME"
 done
 
 
-echo 
-read -r -p "Would you like to install the plugin manager (zplug) for zsh? [y/N] " responsezsh
-responsezsh=${responsezsh,,} # to lower case
-if [[ $responsezsh =~ ^(yes|y)$ ]]; then
+if [ ! -e "$HOME/.zplug" ]; then
+  echo 
+  read -r -p "Would you like to install the plugin manager (zplug) for zsh? [y/N] " responsezsh
+  responsezsh=${responsezsh,,} # to lower case
+  if [[ $responsezsh =~ ^(yes|y)$ ]]; then
     git clone https://github.com/zplug/zplug "$HOME/.zplug"
     echo "Open or reopen zsh to finish installation!"
+  fi
 fi
 
-echo
-read -r -p "Would you like to install the plugin manager (tpm) for tmux? [y/N] " responsetmux
-responsetmux=${responsetmux,,} # to lower case
-if [[ $responsetmux =~ ^(yes|y)$ ]]; then
+if [ ! -e "$HOME/.tmux/plugins/tpm" ]; then
+  echo
+  read -r -p "Would you like to install the plugin manager (tpm) for tmux? [y/N] " responsetmux
+  responsetmux=${responsetmux,,} # to lower case
+  if [[ $responsetmux =~ ^(yes|y)$ ]]; then
     git clone https://github.com/tmux-plugins/tpm "$HOME/.tmux/plugins/tpm"
     echo "Open tmux to finish installation! If you already have a tmux session"
     echo "running, run 'tmux source ~/.tmux.conf'."
+  fi
 fi
     
-echo
-read -r -p "Would you like to install the plugin manager (vim-plug) for vim? [y/N] " responsevim
-responsevim=${responsevim,,} # to lower case
-if [[ $responsevim =~ ^(yes|y)$ ]]; then
+if [ ! -e "$HOME/.vim/autoload/plug.vim" ] || [ ! -e "$HOME/.local/share/nvim/site/autoload/plug.vim" ]; then
+  echo
+  read -r -p "Would you like to install the plugin manager (vim-plug) for vim and nvim? [y/N] " responsevim
+  responsevim=${responsevim,,} # to lower case
+  if [[ $responsevim =~ ^(yes|y)$ ]]; then
     curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
     curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
     echo "Once vim is open, run :PlugInstall to finish installation!"
+  fi
 fi
+
 echo "Done!"
